More work on css/front end layout
Created the input for the image files, styled the box with a honeycomb pattern made with css radial grids. Also used react-particle-js to get the floating pattern in the background. Currently, no functionality has been added. Just trying to get an idea of how the front will be laid out before implementing the code and DB.